[d3d11] Flush immediate context on command list execution
This optimization may help keep the GPU busy in case there's
a large number of draw calls pending at the time a command
list from a deferred context is submitted for execution.

+++ src/d3d11/d3d11_context.h
@@ -648,7 +648,7 @@ namespace dxvk {
     Com<D3D11RasterizerState>   m_defaultRasterizerState;
     
     D3D11ContextState           m_state;
-    uint64_t                    m_drawCount = 0;
+    UINT                        m_drawCount = 0;
     
     void ApplyInputLayout();
